refactor: Direction Enum삭제 폰의 이동규칙을 white,black을 따로 나눠 구현
